cart changes

+++ client/src/components/home.js
@@ -6,7 +6,7 @@ import Navbar from './master/navbar'
 import '../dist/styles/css/home.css'
 import { CartContext } from './master/context/cart'
 import axios from 'axios'
-import {addClass,removeClass } from '../functions/functions'
+import { addClass, removeClass } from '../functions/functions'
 
 export default class Home extends Component {
     static contextType = CartContext
@@ -26,28 +26,32 @@ export default class Home extends Component {
     }
     CalculateSubtotal(CartItems) {
         var total = 0;
-        CartItems.forEach(element => {
-            total = total + parseFloat(element.ProductPrice)
-        });
+        if (CartItems !== undefined) {
+            CartItems.forEach(element => {
+                total = total + parseFloat(element.ProductPrice)
+            });
+        }
         return total
     }
     RenderCartItems = () => {
         this.CartItems = []
         const { CartItems } = this.context
-        CartItems.forEach(element => {
-            this.CartItems.push(
-                <li>
-                    <div className="cart-item">
-                        <img src={element.ProductImage} alt="Cart-Product" className="cart-product-image" />
-                        <div className="cart-item-details">
-                            <h6>{element.ProductName}</h6>
-                            <h6>Rs.{element.ProductPrice}</h6>
-                            <h6 className="Remove-Cart-Item">Remove Item</h6>
+        if (CartItems !== undefined) {
+            CartItems.forEach(element => {
+                this.CartItems.push(
+                    <li>
+                        <div className="cart-item">
+                            <img src={element.ProductImage} alt="Cart-Product" className="cart-product-image" />
+                            <div className="cart-item-details">
+                                <h6>{element.ProductName}</h6>
+                                <h6>Rs.{element.ProductPrice}</h6>
+                                <h6 className="Remove-Cart-Item">Remove Item</h6>
+                            </div>
                         </div>
-                    </div>
-                </li>
-            )
-        })
+                    </li>
+                )
+            })
+        }
         return this.CartItems
     }
     RenderProductCards = () => {
@@ -99,7 +103,7 @@ export default class Home extends Component {
                     {this.RenderProductCards()}
                 </div>
                 <div className="mobile-cart-expanded">
-                <img src={require('../dist/assets/icons/icons8-delete-50.png')} className="Mobile-Cart-Close" alt="close" onClick={() => removeClass(document.getElementsByClassName('mobile-cart-expanded')[0],'show')} />
+                    <img src={require('../dist/assets/icons/icons8-delete-50.png')} className="Mobile-Cart-Close" alt="close" onClick={() => removeClass(document.getElementsByClassName('mobile-cart-expanded')[0], 'show')} />
                     <h2>Cart Items</h2>
                     <div className="ul-container">
                         <ul>
@@ -120,15 +124,14 @@ export default class Home extends Component {
                             <h4>Rs.{this.CalculateSubtotal(CartItems)}</h4>
                         </div>
                         <div className="CheckoutBtn-Container">
-                            <button onClick={() => {this.props.history.push('/checkout')}}>Checkout</button>
+                            <button onClick={() => { this.props.history.push('/checkout') }}>Checkout</button>
                         </div>
                     </div>
                 </div>
-                <div className="mobile-cart" onClick={() => 
-                    {addClass(document.getElementsByClassName('mobile-cart-expanded')[0],'show')}
+                <div className="mobile-cart" onClick={() => { addClass(document.getElementsByClassName('mobile-cart-expanded')[0], 'show') }
                 }>
                     <img src={require('../dist/assets/icons/icons8-shopping-cart-48.png')} alt="cart"></img>
-                    <div className="mobile-cart-noti" id="mobile-cart-noti">{CartItems.length}</div>
+                    <div className="mobile-cart-noti" id="mobile-cart-noti">{CartItems !== undefined ? CartItems.length : 0}</div>
                 </div>
             </div>
         )
